more changes

+++ sql/expression/function/srid.go
@@ -34,8 +34,8 @@ var _ sql.FunctionExpression = (*SRID)(nil)
 var ErrInvalidSRID = errors.NewKind("There's no spatial reference with SRID %d")
 
 const (
-	CartesianSRID  = 0
-	GeoSpatialSRID = 4326
+	CartesianSRID  = uint32(0)
+	GeoSpatialSRID = uint32(4326)
 )
 
 // NewSRID creates a new STX expression.
